+++ backend/tests/test_service_org_unit.py
@@ -1,25 +1,24 @@
 # SPDX-FileCopyrightText: 2018-2020 Magenta ApS
 # SPDX-License-Identifier: MPL-2.0
 
-import json
-
 import freezegun
-import requests_mock
-
-from mock import patch
-from mora import lora
+from mock import call, patch
+from mora import lora, mapping
 from mora.async_util import async_to_sync
 from mora.exceptions import HTTPException
 from mora.handler.impl.association import AssociationReader
 from mora.service.orgunit import UnitDetails, _get_count_related, get_one_orgunit
+from mora.triggers import Trigger
+from mora.triggers.internal.http_trigger import HTTPTriggerException, register
+from os2mo_http_trigger_protocol import MOTriggerRegister
 from tests import util
 
 
 class TestAddressLookup(util.TestCase):
-    @freezegun.freeze_time('2018-03-15')
+    @freezegun.freeze_time("2018-03-15")
     @util.MockAioresponses()
     def test_unit_past(self, mock):
-        unitid = 'ef04b6ba-8ba7-4a25-95e3-774f38e5d9bc'
+        unitid = "ef04b6ba-8ba7-4a25-95e3-774f38e5d9bc"
 
         reg = {
             "attributter": {
@@ -31,15 +30,15 @@ class TestAddressLookup(util.TestCase):
                             "from": "1993-01-01 00:00:00+01",
                             "from_included": True,
                             "to": "infinity",
-                            "to_included": False
-                        }
+                            "to_included": False,
+                        },
                     }
                 ]
             },
             "brugerref": "42c432e8-9c4a-11e6-9f62-873cf34a735f",
             "fratidspunkt": {
                 "graenseindikator": True,
-                "tidsstempeldatotid": "2018-03-09T14:38:45.310653+01:00"
+                "tidsstempeldatotid": "2018-03-09T14:38:45.310653+01:00",
             },
             "livscykluskode": "Rettet",
             "relationer": {
@@ -51,8 +50,8 @@ class TestAddressLookup(util.TestCase):
                             "from": "1993-01-01 00:00:00+01",
                             "from_included": True,
                             "to": "infinity",
-                            "to_included": False
-                        }
+                            "to_included": False,
+                        },
                     },
                     {
                         "objekttype": "80764a2f-6a7b-492c-92d9-96d24ac845ea",
@@ -61,9 +60,9 @@ class TestAddressLookup(util.TestCase):
                             "from": "1993-01-01 00:00:00+01",
                             "from_included": True,
                             "to": "infinity",
-                            "to_included": False
-                        }
-                    }
+                            "to_included": False,
+                        },
+                    },
                 ],
                 "enhedstype": [
                     {
@@ -72,8 +71,8 @@ class TestAddressLookup(util.TestCase):
                             "from": "1993-01-01 00:00:00+01",
                             "from_included": True,
                             "to": "infinity",
-                            "to_included": False
-                        }
+                            "to_included": False,
+                        },
                     }
                 ],
                 "overordnet": [
@@ -83,8 +82,8 @@ class TestAddressLookup(util.TestCase):
                             "from": "1993-01-01 00:00:00+01",
                             "from_included": True,
                             "to": "infinity",
-                            "to_included": False
-                        }
+                            "to_included": False,
+                        },
                     }
                 ],
                 "tilhoerer": [
@@ -94,8 +93,8 @@ class TestAddressLookup(util.TestCase):
                             "from": "1993-01-01 00:00:00+01",
                             "from_included": True,
                             "to": "infinity",
-                            "to_included": False
-                        }
+                            "to_included": False,
+                        },
                     }
                 ],
                 "tilknyttedeenheder": [
@@ -104,10 +103,10 @@ class TestAddressLookup(util.TestCase):
                             "from": "1993-01-01 00:00:00+01",
                             "from_included": True,
                             "to": "infinity",
-                            "to_included": False
+                            "to_included": False,
                         }
                     }
-                ]
+                ],
             },
             "tilstande": {
                 "organisationenhedgyldighed": [
@@ -117,8 +116,8 @@ class TestAddressLookup(util.TestCase):
                             "from": "1993-01-01 00:00:00+01",
                             "from_included": True,
                             "to": "infinity",
-                            "to_included": False
-                        }
+                            "to_included": False,
+                        },
                     },
                     {
                         "gyldighed": "Inaktiv",
@@ -126,22 +125,19 @@ class TestAddressLookup(util.TestCase):
                             "from": "-infinity",
                             "from_included": True,
                             "to": "1993-01-01 00:00:00+01",
-                            "to_included": False
-                        }
-                    }
+                            "to_included": False,
+                        },
+                    },
                 ]
             },
-            "tiltidspunkt": {
-                "tidsstempeldatotid": "infinity"
-            }
+            "tiltidspunkt": {"tidsstempeldatotid": "infinity"},
         }
 
         mock.get(
-            'http://mox/organisation/organisationenhed'
-            '?uuid=' + unitid +
-            '&virkningtil=2018-03-15T00%3A00%3A00%2B01%3A00'
-            '&virkningfra=-infinity'
-            '&konsolider=True',
+            "http://mox/organisation/organisationenhed"
+            "?uuid=" + unitid + "&virkningtil=2018-03-15T00%3A00%3A00%2B01%3A00"
+            "&virkningfra=-infinity"
+            "&konsolider=True",
             payload={
                 "results": [
                     [
@@ -149,93 +145,160 @@ class TestAddressLookup(util.TestCase):
                             "id": "ef04b6ba-8ba7-4a25-95e3-774f38e5d9bc",
                             "registreringer": [
                                 reg,
-                            ]
+                            ],
                         }
                     ]
                 ]
-            }
+            },
         )
 
         self.assertRequestResponse(
-            '/service/ou/' + unitid + '/details/org_unit?validity=past',
+            "/service/ou/" + unitid + "/details/org_unit?validity=past",
             [],
         )
 
 
 class TestTriggerExternalIntegration(util.TestCase):
-    @patch('mora.service.orgunit.get_one_orgunit')
+    @patch("mora.service.orgunit.get_one_orgunit")
     def test_returns_404_on_unknown_unit(self, mock):
         mock.return_value = {}
 
         r = self.assertRequest(
-            '/service/ou/44c86c7a-cfe0-447e-9706-33821b5721a4/trigger-external',
-            status_code=404
+            "/service/ou/44c86c7a-cfe0-447e-9706-33821b5721a4/refresh", status_code=404
         )
-        self.assertIn('NOT_FOUND', r.get('error_key'))
+        self.assertIn("NOT_FOUND", r.get("error_key"))
 
-    @util.override_config({"external_integration": {"org_unit": "http://whatever/"}})
-    @patch('mora.service.orgunit.get_one_orgunit')
-    @requests_mock.Mocker()
-    def test_returns_integration_error_on_wrong_status(self, mock, r_mock):
-        mock.return_value = {'whatever': 123}
+    @util.override_config(
+        {
+            "triggers": {
+                "http_trigger": {
+                    "enabled": True,
+                    "http_endpoints": ["http://whatever"],
+                }
+            }
+        }
+    )
+    @patch("mora.triggers.internal.http_trigger.fetch_endpoint_trigger")
+    @patch("mora.triggers.internal.http_trigger.http_sender")
+    @patch("mora.service.orgunit.get_one_orgunit")
+    def test_returns_integration_error_on_wrong_status(
+        self, mock, t_sender_mock, t_fetch_mock
+    ):
+        t_fetch_mock.return_value = [
+            MOTriggerRegister(
+                **{
+                    "event_type": mapping.EventType.ON_BEFORE,
+                    "request_type": mapping.RequestType.REFRESH,
+                    "role_type": "org_unit",
+                    "url": "/triggers/ou/refresh",
+                }
+            )
+        ]
+        Trigger.registry = {}
+        register(None)
+        t_fetch_mock.assert_called()
 
-        error_msg = "Something bad happened"
+        error_msg = "Something horrible happened"
 
-        r_mock.post(
-            (
-                "http://whatever/"
-            ),
-            text=error_msg,
-            status_code=500
-        )
+        def side_effect(trigger_url, trigger_dict, timeout):
+            raise HTTPTriggerException(error_msg)
 
+        t_sender_mock.side_effect = side_effect
+
+        mock.return_value = {"whatever": 123}
         r = self.assertRequest(
-            '/service/ou/44c86c7a-cfe0-447e-9706-33821b5721a4/trigger-external',
-            status_code=400
+            "/service/ou/44c86c7a-cfe0-447e-9706-33821b5721a4/refresh", status_code=400
         )
-        self.assertIn('INTEGRATION_ERROR', r.get('error_key'))
-        self.assertIn(error_msg, r.get('description'))
+        self.assertIn("INTEGRATION_ERROR", r.get("error_key"))
+        self.assertIn(error_msg, r.get("description"))
 
-    @util.override_config({"external_integration": {"org_unit": "http://whatever/"}})
-    @patch('mora.service.orgunit.get_one_orgunit')
-    @requests_mock.Mocker()
-    def test_returns_message_on_success(self, mock, r_mock):
-        mock.return_value = {'whatever': 123}
+        t_sender_mock.assert_called_with(
+            "http://whatever/triggers/ou/refresh",
+            {
+                "request_type": mapping.RequestType.REFRESH,
+                "request": {"uuid": "44c86c7a-cfe0-447e-9706-33821b5721a4"},
+                "role_type": "org_unit",
+                "event_type": mapping.EventType.ON_BEFORE,
+                "org_unit_uuid": "44c86c7a-cfe0-447e-9706-33821b5721a4",
+                "uuid": "44c86c7a-cfe0-447e-9706-33821b5721a4",
+            },
+            timeout=5,
+        )
+
+    @util.override_config(
+        {
+            "triggers": {
+                "http_trigger": {
+                    "enabled": True,
+                    "http_endpoints": ["http://whatever"],
+                }
+            }
+        }
+    )
+    @patch("mora.triggers.internal.http_trigger.fetch_endpoint_trigger")
+    @patch(
+        "mora.triggers.internal.http_trigger.http_sender", new_callable=util.CopyingMock
+    )
+    @patch("mora.service.orgunit.get_one_orgunit")
+    def test_returns_message_on_success(self, mock, t_sender_mock, t_fetch_mock):
+        t_fetch_mock.return_value = [
+            MOTriggerRegister(
+                **{
+                    "event_type": mapping.EventType.ON_BEFORE,
+                    "request_type": mapping.RequestType.REFRESH,
+                    "role_type": "org_unit",
+                    "url": "/triggers/ou/refresh",
+                }
+            )
+        ]
+        Trigger.registry = {}
+        register(None)
+        t_fetch_mock.assert_called()
 
         response_msg = "Something good happened"
+        t_sender_mock.return_value = response_msg
 
-        r_mock.post(
-            (
-                "http://whatever/"
-            ),
-            text=json.dumps({"output": response_msg}),
-            status_code=201
-        )
-
+        mock.return_value = {"whatever": 123}
         r = self.assertRequest(
-            '/service/ou/44c86c7a-cfe0-447e-9706-33821b5721a4/trigger-external',
+            "/service/ou/44c86c7a-cfe0-447e-9706-33821b5721a4/refresh"
         )
-        self.assertIn(response_msg, r['message'])
-        self.assertEqual(201, r.get('status_code'))
+        self.assertEqual(response_msg, r["message"])
+
+        t_sender_mock.assert_has_calls(
+            [
+                call(
+                    "http://whatever/triggers/ou/refresh",
+                    {
+                        "request_type": mapping.RequestType.REFRESH,
+                        "request": {"uuid": "44c86c7a-cfe0-447e-9706-33821b5721a4"},
+                        "role_type": "org_unit",
+                        "event_type": mapping.EventType.ON_BEFORE,  # WHAT THE FUCK
+                        "org_unit_uuid": "44c86c7a-cfe0-447e-9706-33821b5721a4",
+                        "uuid": "44c86c7a-cfe0-447e-9706-33821b5721a4",
+                    },
+                    timeout=5,
+                )
+            ]
+        )
 
 
 class TestGetOneOrgUnit(util.LoRATestCase):
     def setUp(self):
         self.load_sample_structures(minimal=True)
         self._connector = lora.Connector(
-            virkningfra='-infinity', virkningtil='infinity'
+            virkningfra="-infinity", virkningtil="infinity"
         )
-        self._orgunit_uuid = '2874e1dc-85e6-4269-823a-e1125484dfd3'
+        self._orgunit_uuid = "2874e1dc-85e6-4269-823a-e1125484dfd3"
 
     def test_details_nchildren(self):
         self._assert_orgunit_keys(
-            {'uuid', 'name', 'user_key', 'validity', 'child_count'},
+            {"uuid", "name", "user_key", "validity", "child_count"},
             details=UnitDetails.NCHILDREN,
         )
 
     def test_details_path(self):
         self._assert_orgunit_keys(
-            {'uuid', 'name', 'user_key', 'validity', 'location'},
+            {"uuid", "name", "user_key", "validity", "location"},
             details=UnitDetails.PATH,
         )
 
@@ -243,9 +306,9 @@ class TestGetOneOrgUnit(util.LoRATestCase):
         result = async_to_sync(get_one_orgunit)(
             self._connector,
             self._orgunit_uuid,
-            count_related={'association': AssociationReader},
+            count_related={"association": AssociationReader},
         )
-        self.assertIn('association_count', result)
+        self.assertIn("association_count", result)
 
     def _assert_orgunit_keys(self, expected_keys, **kwargs):
         orgunit = async_to_sync(get_one_orgunit)(
@@ -257,22 +320,22 @@ class TestGetOneOrgUnit(util.LoRATestCase):
 class TestGetCountRelated(util.TestCase):
     def setUp(self):
         super().setUp()
-        self._simple = {'association'}
-        self._multiple = {'association', 'engagement'}
+        self._simple = {"association"}
+        self._multiple = {"association", "engagement"}
 
     def test_valid_name(self):
-        with self.app.test_request_context('?count=association'):
+        with self.app.test_request_context("?count=association"):
             self.assertSetEqual(self._simple, _get_count_related())
 
     def test_valid_name_repeated(self):
-        with self.app.test_request_context('?count=association&count=association'):
+        with self.app.test_request_context("?count=association&count=association"):
             self.assertSetEqual(self._simple, _get_count_related())
 
     def test_multiple_valid_names(self):
-        with self.app.test_request_context('?count=association&count=engagement'):
+        with self.app.test_request_context("?count=association&count=engagement"):
             self.assertSetEqual(self._multiple, _get_count_related())
 
     def test_invalid_name(self):
-        with self.app.test_request_context('?count=association&count=foobar'):
+        with self.app.test_request_context("?count=association&count=foobar"):
             with self.assertRaises(HTTPException):
                 _get_count_related()