+++ NEWS.rst
@@ -13,6 +13,7 @@ New features
 * #42112: Allow UI to set and change engagement ID (bvn)
 * #42033: Made extension_X fields on engagements UI-wise configurable in conf_db
 * #41973: Allowed addresses pointing to engagements
+* #41894: Added refresh trigger event
 
 Version 1.15.0, 2021-03-08
 ==========================